Add syntactic sugar for PartialFunction [dotc]
